Fixed GetDeltaTime bug on linux.
The GetDeltaTime was returning a 32bit value and the linux implementation of get_a_ticks returns a long int so the next frame time was never hit.

+++ engine/include/GameEngine.h
@@ -51,7 +51,7 @@ public:
     void PlayOne();
     void PlayOneUnlocked();
 
-    int GetDeltaTime();
+    long GetDeltaTime();
 
     // TModule has to inherit from GameModule
     template <typename TModule = GameModule>
